+++ engines/suppliers/app/models/suppliers/group.rb
@@ -9,14 +9,5 @@ module Suppliers
     has_many :producers,
       class_name: 'Suppliers::Producer'.freeze,
       through: :suppliers
-
-    has_many :memberships, foreign_key: :basic_resource_group_id
-
-    def admin?(user)
-      memberships.where(
-        role: Membership::ROLES[:admin],
-        user_id: user.id
-      ).any?
-    end
   end
 end
